fix: chipperId now Long instead of Integer

+++ src/main/java/com/dripchip/api/entity/Animal.java
@@ -27,7 +27,7 @@ public class Animal {
     private Gender gender;
     private LifeStatus lifeStatus;
     private LocalDateTime chippingDateTime;
-    private Integer chipperId;
+    private Long chipperId;
     private Long chippingLocationId;
     private List<Long> visitedLocations;
     private LocalDateTime deathDateTime;
